refactor(cli): consolidate most MainWorker related code to the same place (#15459)
